requirements.txt
azure-identity
openai
python-dotenv
pydantic
rich
azure-ai-inference
dotenv-azd
aiohttp
autogen-agentchat
autogen-ext[openai]
azure-ai-inference==1.0.0b9
openai-agents
semantic-kernel
langgraph
langchain_openai
pydantic-ai
llama-index
llama-index-llms-azure-openai
llama-index-llms-openai-like
llama-index-embeddings-azure-openai
smolagents
